Implemented a method for restricting the expansion of HMM bands to allow for marginal hits. Previously all MP, ML, MR state bands were expanded to allow for all possible marginal hits. Now only certain state bands are expanded. A state v that roots the subtree spanning lpos..rpos consensus positions is expanded based on occupancy of HMM nodes that model lpos/rpos (probability the match or delete state of the HMM node is visited) from a Forward/Backward/Posterior parse of the target sequence.
For example if occ[lpos] = 0.9999 and occ[rpos] = 0.9999, this means
the HMM is almost certain lpos and rpos are part of the optimal
alignment, so we don't have to worry about looking for marginal
alignments that span lpos..rpos. Conversely, if occ[lpos] and
occ[rpos] are very close to 0. then we don't have to worry about a
joint (non-marginal) alignment that spans lpos..rpos. Other exclusions
specific to L, R and T matrices are also possible. During HMM banded
trCYK, the J, L, R, T matrices can be skipped for states where the
corresponding joint or marginal alignment is unlikely based on the HMM
occupancy values. Currently, FastTrCYKScanHB() does not skip any of
these states. FastTrCYKScanHBSkipSlow() does skip them, but only by
checking if they should be skipped in the deepest part of the
recursion (very inefficient). It's only roughly 20% faster than
FastTrCYKScanHB() at best. I'll try to speed it up soon.  


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3677 db3e70e4-12e1-0310-b23f-81a6e3d09ba7