Simplified HMM banded trCYK band calculation and alignment. Current (new) strategy is to predict four consensus positions, stored in the CP9Bands_t data structure based on the HMM posteriors: sp1: first 'possible' consensus position used in parse (min k for      which occ[k] > thresh1 (typically 0.01)) sp2: first 'probable' consensus position used in parse (min k for      which occ[k] > thresh2 (typically 0.99)) ep1: final 'possible' consensus position used in parse (max k for      which occ[k] > thresh1 (typically 0.01)) ep2: final 'probable' consensus position used in parse (max k for      which occ[k] > thresh2 (typically 0.99))
Where occ[k] is the probability that HMM node k is occupied, i.e. the
summed probability that node k's match or delete state is used for any
residue in the sequence.

Then, for each CM state v, that roots a subtree spanning lpos to rpos
in consensus positions, determine if Joint, Left, Right, Terminal
marginal alignments are possible for v based on where lpos and rpos
fall relative to sp1, sp2, ep1, ep2. For example, Joint alignment is
only allowed if lpos >= sp1 and rpos <= ep1. Left alignment is only
allowed if sp1 <= lpos <= ep1 and rpos > ep2. Other rules dictate
Right and Terminal alignment possibilities. 

HMM bands are calculated as in the non-truncated case and expanded for
states that allow L, R, and T alignment. For L/T alignment, j bands are
expanded to allow alignment of any residue that may align to positions
ep2 or ep1. For R/T alignment, i bands are expanded to allow alignment
of any residue that may align to positions sp2 or sp1.

There is a potentially problematic case if no occ[k] exceeds thresh2
and/or thresh1, which causes J,L,R,T alignment to be possible for all
states, and HMM bands to be expanded to allow all possible residues to
align to any state. This happens commonly when aligning random
sequences for example, and alignment is still possible, but the
truncated HMM banded matrix may be too big. In this case, the
currently implemented strategy is to raise thresh1 and lower thresh2
concurrently with raising the HMM band probability threshold tau until
the resulting matrix is small enough.

Currently, code for the previous, more complicated way of calculated
bands for marginal alignments is included. It will be removed soon.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3681 db3e70e4-12e1-0310-b23f-81a6e3d09ba7