cm_TrInside() and cm_TrOutside() are now consistent (based on my current testing on small models) and the resulting {J,L,R,T}alpha and {J,L,R,T}beta matrices can be combined into a valid posterior matrix using cm_TrPosterior(). The cm_TrCheckPosterior() function checks this matrix to make sure that each residue has a summed probability of 1.0 of being emitted from some state in the model.
The optimal 'Inside' score is currently calculated as the maximum
of the summed log odds score of: 

(a) all valid J parsetrees
(b) all valid L parsetrees
(c) all valid R parsetrees
(d) all valid T parsetrees

That is, it isn't the sum of all possible parsetrees in all modes. To
get a set of Outside matrices from which valid Posteriors can be
calculated, we need to restrict possible parsetree modes based on
which of (a), (b), (c) or (d) defined the Inside score. If:

(a) only allow J      parsetrees in the Outside Beta matrices
(b) only allow J or L parsetrees in the Outside Beta matrices
(c) only allow J or R parsetrees in the Outside Beta matrices
(d) allow all parsetrees in the Outside Beta matrices

If we don't restrict the possible parsetrees then we can get
[v][j][d] cells for which 
{J,L,R,T}alpha[v][j][d] + {J,L,R,T}beta[v][j][d] exceeds the
Inside score, which we don't want. Fortunately, in Outside it's simple
to restrict the mode of possible parsetrees during initialization of
cells for which j == L and d == L (for sequence length L), by setting
cells for disallowed modes to IMPOSSIBLE.

Restricting modes also allows an easy speedup and memory savings of
the DP algorithms, by skipping updates of cells in disallowed modes. I
haven't implemented this yet though.

We should be able to restrict modes when doing cm_TrCYKAlign() or
cm_TrInsideAlign() as well if we're in the search pipeline because the
first pass TrCYK/TrInside scanner will give us the optimal mode.

ref: ELN3 notebook p8-11
ref: ~nawrockie/notebook/11_0816_inf_trcyk_banded/00LOG



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3706 db3e70e4-12e1-0310-b23f-81a6e3d09ba7