Most changes are in cm_dpalign.c, where I did general code clean-up and modified some function prototypes.
The main motivation was to force all alignment functions to take only
'L' the sequence length instead of 'i0' and 'j0' start and end
points. Using i0 and j0 was left over from the divide and conquer
(D&C) small memory functions upon the cm_dpalign.c functions were
originally based, but it no longer applies. The D&C functions are
designed to be called recursively on subsequences, where i0 != 1 and
j0 != L, but all of the alignment functions in cm_dpalign.c are
designed to be called once on the full sequence 1..L. Using both i0
and j0 was misleading and probably wouldn't even work in some cases if
i0 != 1. Making this change simplifies the code a small amount
throughout all of the DP functions because we know that matrix
position j corresponds to sequence position j, whereas before it could
be offset by i0-1.  While I was doing this I took the opportunity to
generally clean up the code and make the various functions, many of
which are very similar (CYK and Inside functions for example) as
similar as possible.

This introduces a bunch of minor, trivial code changes to the DP
recursions in cm_dpalign.c. Everything 'should' still work, but any
typo I may have made could certainly have broken something. Future
testing should find any of these. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3709 db3e70e4-12e1-0310-b23f-81a6e3d09ba7