Added HMM banded truncated outside, optimal accuracy, posterior, and 'EmitterPosterior' functions.
Optimal accuracy tracebacks are currently flawed in that transitions
from silent inserts (ILs in Right marginal mode and IRs in Left
marginal mode) are preferred over match states simply because those
inserts are inspected before the match states in the DP recursion and
both have identical scores. This leads to unnecessary states in the
optimally accurate parsetree. I think the only reasonable way to fix
this is to create new ordering of possible child states to transition
to for Left and Right marginal mode updates in the optimal accuracy DP
recursion. I'll attempt this soon.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3716 db3e70e4-12e1-0310-b23f-81a6e3d09ba7