cm_modelconfig.c:
Changed default behavior for how CP9 HMMs are locally
configured. Previously, they were configured to be as similar to the
locally configured CM as possible, with 1.-cm->pbegin probability of
starting with the first match state, and 1.-cm->pend probability of
ending with the final match state. Also, the I_0 and I_M (insert state
of first and last nodes) were unreachable because the CM states they
correspond to ROOT_IL and ROOT_IR are unreachable in a locally
configured CM.

Now, to better handle HMM banding for truncated alignment, all HMM
start/stops are equiprobable, and I_0 and I_M are reachable, because
they are in truncated local HMM alignment. In search I_0 and I_M will
never be used because they will never increase the bit score, but in
'cmalign' they should be reachable. Since the plan is to use truncated
alignment as default in cmalign, this change was necessary. I added a
flag in structs.h, CM_CONFIG_V1P0_HMM_LOCAL to allow reverting to the
old way, although no programs currently have options for this. I plan
to benchmark the effect of this on Diana's benchmark from the trCYK
paper. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3736 db3e70e4-12e1-0310-b23f-81a6e3d09ba7