Modified how truncated alignments are displayed in cmsearch/cmscan output via the CM_ALIDISPLAY object.
Previously, all truncated ends were shown, treated like local ends
(EL) for output, but this often included entire subtrees rooted at S
states being deleted thanks to the truncation. I thought this was
confusing and now only up to two additional EL-like insertions are
added for a truncated alignment. One indicating the number of
consensus model positions truncated at the 5' end at the beginning of
the alignment (if aln is R or T mode), and the other showing the
number of consensus model positions truncated at the 3' end at the end
of the alignment. Also, basepairs for which the other half has been
truncated-away are annotated with '?' in the 'NC' annotation line
indicating it is unknown if these pairs are negative-scoring
non-canonicals or not. I'm not sure if I'll keep this, but one nice
side effect of this is it helps orient the user as to how the
alignment is truncated, and points out that the user should not
attempt to look for the other half of such pairs in the alignment.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3743 db3e70e4-12e1-0310-b23f-81a6e3d09ba7