Serious changes introduced to how truncated hits are detected and aligned. Now, by default:
 - the first residue of a target sequence must exist in R marginal
   alignments (truncation of 5' end of the RNA)
 - the final residue of a target sequence must exist in L marginal
   alignments (truncation of 3' end of the RNA)
 - the first and final residues of a target sequence must exist in T
   marginal alignments (truncation of 5' and 3' end of the RNA)

This change was introduced to reduce FPs in regions where truncated
hits are allowed (i.e. the first cm->W and final cm->W residues of
each target sequence) and because I believe it is consistent with the
assumptions underlying the truncated paradigm, that residues outside
the truncated hit are missing data. (For example, allowing a R
truncated hit to begin at position x != 1 ignores the fact that
residues 1..x-1 are not actually missing.)

Current strategy for enforcing first and/or final residues are
contained within truncated hits involves specially configured p7 HMMs
for L, R, and T modes, supported in a new file:
cm_p7_modelconfig_trunc.c. I am still experimenting with this. 

I plan on continuing to make changes to this new strategy in future
revisions, but this is a good commit point. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3757 db3e70e4-12e1-0310-b23f-81a6e3d09ba7