Simplified CP9 HMM dynamic programming functions in cp9_dp.c thanks to their narrower scope. They are no longer used in 'search' mode, p7 HMMs and HMMER routines are now used. They are still used for deriving bands for CM search/alignment though. The changes to those function prototypes is responsible for many of the modified files here.
Still tweaking truncated search/alignment in the pipeline. I've fixed
a few problems from the previous revision; performance on Diana's
benchmark is improved.

I'm about to overhaul how cmsearch/cmscan identify the sequence ends,
which are re-searched for possible truncated hits. Currently, each of
the main loops in cmsearch/cmscan have to handle this (serial_loop,
pipeline_thread, and mpi_worker) and it is messy. I'm going to try to
shift this responsibility to the more general cm_Pipeline() function,
which should be cleaner.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3760 db3e70e4-12e1-0310-b23f-81a6e3d09ba7