Overhauled how 5' and 3' ends of sequences are identified and re-searched allowing for truncated hits. Previously, cmsearch and cmscan handled this by copying the W terminal residues and calling cm_Pipeline() with the terminus as a new sequence. This made the otherwise straightforward serial, threaded and MPI loops significantly more complex than they should be and required duplication of the relevant code amongst those functions. Now, cm_Pipeline() handles identifying, extracting and researching the terminal sequence.
Unfortunately, in non-MPI cmsearch we don't always know the full
length of a target sequence while we are searching it, because long
sequences are input in overlapping 100Kb chunks. So we re-search the
3' end of all 100 Kb chunks, just in case each one turns out to be the
final one. This is wasteful, but not terribly, for each such 100Kb
chunk we re-search W residues, which on average should be about a 0.2%
duplication of effort.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3764 db3e70e4-12e1-0310-b23f-81a6e3d09ba7