Two significant changes:
1. Revised how truncated alignment interacts with local begins. 

Added cm->trbegin and cm->trbeginsc arrays to the CM_t data
structure. These are analogous to cm->begin and cm->beginsc arrays but
allow local begins into insert states as well as main match
states. cm->trbegin probabilities are determined similarly to
cm->begin probabilities but also use the relative occupancy of inserts
versus match states. Begins into inserts are necessary because we
enforce the first/final residue of a target be included in any 5'/3'
truncated hit, and we need to be able to handle it if that residue is
an insert.

In global mode, a truncated hit must be rooted at state 0 or an
internal BIF_B state in the case of a Terminal (T) alignment. I
believe any such T alignment should be given a penalty of the summed
score of transitions from ROOT_S to the relevant BIF_B, but this isn't
implemented yet.

These changes were introduced to make scoring of truncated and
standard hits as consistent as possible, so I can reasonably use
standard E-values for truncated hits. 

2. Added a new CP9 HMM to the CM data structure for dealing with
   truncated alignment for a globally configured CM. 

This change is responsible for many of the modified files.  It
required renaming cm->cp9 to cm->cp9loc and adding a new CP9 HMM
called cm->cp9glb. The former is always locally configured and the
latter always globally. In glocal search mode, cm->cp9loc is used to
determine bands for truncated hits, and cm->cp9glb is used for
standard hits.  Many functions were previously passed a cm and were
using cm->cp9. These have been changed to take a cp9 directly and the
caller has to know if it should be cm->cp9loc or cm->cp9glb.

Just before I finished implementing this I thought of a better way,
which requires three additional cp9 HMM's one for each truncation
mode. I will work on this next, but wanted to commit current code so
I could revert if necessary.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3774 db3e70e4-12e1-0310-b23f-81a6e3d09ba7