Rewrote cmalign from scratch. It comes in as 'x-cmalign.c', 'cmalign.c' is the old version. I'll switch the name soon.
Main motivation was to get threads working, and I wanted to simplify
the code as well. A secondary goal was to get rid of the old
DispatchAlignments() function. It is now replaced by a
'process_align_workunit()' local to x-cmalign.c (but this will have to
be made external soon so cmbuild can call it when --refine is
used). Serial and threaded are working, though there are a few options
still to add. MPI is not yet implemented. 

Additional changes relative to old (v1.0.2) cmalign: 
o truncated alignment is implemented, and is default.
o --mapali, based on hmmalign's option of the same name, comes in as a
  replacement to --withali. New implementation is much simpler because
  the new CM file format stores an alignment checksum and alignment
  map from the input alignment. --mapstr makes the input alignment
  structure be used (including pseudoknots), this replaces
  --withpknots. 
o --sub is still an option even though truncated alignment could
  replace it. I wanted to leave it in though to allow an analysis of
  the relative efficacy of each approach.
o Memory-saving mode is used, but only if necessary. Alignments are
  output in interleaved Stockholm if and only if all target sequences
  combined less than 10 Mb (this should limit total size of all
  parsetrees to about 250 Mb). If this size is exceeded, then each 10
  Mb chunk is aligned separately and output to a tmp file. After all
  sequences are aligned, the alignments in the tmp file are merged and
  the final alignment of all sequences is output to the normal output
  file in 1-line Pfam Stockholm format.

Unrelated change, that is responsible for many of the modified files
in this commit:
o Removed cmERRBUFSIZE, replaced with eslERRBUFSIZE. cmERRBUFSIZE was
  meant to allow for longer error messages, but it was not able to do
  that in fact because the ESL_FAIL, ESL_XFAIL macros allow a maximum
  length buffer of eslERRBUFSIZE anyway.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3800 db3e70e4-12e1-0310-b23f-81a6e3d09ba7