Non-MPI cmsearch no longer uses an SSI index if one is available. Previously, if available, SSI was used to determine the database size and the total length of each sequence as it was read and searched. If SSI was not available, the entire sequence file was read with repeated calls to esl_sqio_ReadInfo() prior to doing any searches to obtain the total size and each sequence's length. Now, the entire file is read initially regardless of whether an SSI index exists or not.
The behavior was changed because using SSI was only slightly faster in
most test cases and significant slower (~2X overall search time) in
some (when the sequence file had many sequences (~millions)). SSI
doesn't help much here because we need to read the entire file as we
search it anyhow (it would dramatically help if we were only searching
a few sequences in the file, for example). Some advantages of dropping
the use of SSI in non-MPI cmsearch is we've removed some code,
eliminated a branch point, and don't have to rely on the SSI index
being current.

MPI still requires SSI indexing. The reasoning there is that we may be
using hundreds of processors to search and the step of determining
database size must be done prior to the search with a single
processor, so it can be a bottleneck.

As the number of cores on a typical machine increases, we may want to
rethink this, and I may want to refer to the code in the previous
revision which used SSI in non-MPI mode. A better solution would be to
use a file format which puts the sequence length info at the beginning
of the file for quick retrieval.

xref: ~nawrockie/notebook/12_0221_inf_pipeline_finalize/00LOG


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3877 db3e70e4-12e1-0310-b23f-81a6e3d09ba7