Added what is intended to be the final option to any of the v1.1 programs: --wcx <x> to cmsearch and cmscan. This sets the W parameter (max expected hit length) of all models to <x> * clen (model consensus length), ignoring the W from the CM file, which is calculated by cmbuild using QDB and beta=1E-7 (by default).
The motivation for this was my vague recollection of a user telling me
they wanted to be able to set W on the command line. I imagine it
might be useful when trying to find hits that may include large
introns or other inserts. It's likely the default pipeline will find
these anyway if they're good hits, since W really only affects how
windows are defined around hits and multiple MSV hits would be merged
into one big window (possibly >> W) around all of them.  But homologs
with one or few MSV hits might be lost if they're >> W. This option
therefore is an expert one solely included to appease whoever it was
that requested it (maybe Franz Lang??).

<x> from --wcx <x> can't be less than 1.25, because we don't want
users to use W smaller than the model length as that would cause
problems in the filter pipeline, mainly in the glocal Forward stages,
which require a hit to the full model.

--wcx is necessarily incompatible with --nohmm, --qdb or --fqdb all of
which use QDBs at one or more CM stages. If QDBs are used, W is
defined as the max width of a QD band, so trying to set it on the
command line doesn't make sense.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3964 db3e70e4-12e1-0310-b23f-81a6e3d09ba7