Added 1 important line to cmsearch/cmscan. A check in MPI mode if only 1 processor is running after MPI_Init() then we die with cm_Fail(). We'll have no workers, so we can't run. Most likely cause is if user fails to execute 'mpirun' and just does 'cmsearch --mpi'. This check was in cmcalibrate and cmalign, but I hadn't added it to cmsearch/cmscan since rebuilding them from hmmsearch/hmmscan.
I don't think a similar check is in hmmsearch or hmmscan. Should it
be? Without such a check, the MPI process will just run indefinitely.

I'm not sure if a cm_Fail() is kosher, but it seems to get the job
done (it kills the lone process that's running - there's no need to
worry about detached and confused workers).


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3972 db3e70e4-12e1-0310-b23f-81a6e3d09ba7