Added ability to run a nhmmer-like HMM-only pipeline in cmsearch/cmscan.
By default, any model with 0 basepairs will use this
pipeline variant instead of the standard CM pipeline. 
It is also used (for all models) if --hmmonly is used.
This is true in both cmsearch and cmscan.

The motivation for this came out of the Xfam 2012 meeting. There was a
request to use HMMs for structureless CMs in Rfam, while not needing
Rfam curators to switch between nhmmer and Infernal, and not requiring
them to release a CM file for some families and a HMM file for
others. 

The hmmonly pipeline in Infernal is very similar to the one used by
nhmmer. It uses the same filters and filter thresholds. Bit scores are
often identical with nhmmer, though not always, due to subtle
differences in the pipelines. 

The --hmmF1, --hmmF2, --hmmF3, --hmmnobias, --hmmnonull2, and --hmmmax
options control the HMM-only pipeline (independently of the options
that control the CM pipeline).

Output has been modified due to this change. A hit alignment is now
annotated with whether it is to a CM or an HMM (cmscan results can
have these interleaved per query sequence). NC annotation (negative
scoring noncanonicals) is absent from HMM hits, since they'll normally
have 0 basepairs, and even if they do have basepairs (--hmmonly) the
paired positions won't have been scored together. 

Updated manpages and testsuite for this change.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3984 db3e70e4-12e1-0310-b23f-81a6e3d09ba7