More changes due to the introduction of the HMM-only pipeline variant in r3984.
o New prior for zero-bp models in prior.c:Prior_Create().
  Added code copied from p7_prior.c:p7_prior_CreateNucleic() that uses
  the nhmmer default prior for models with 0 basepairs. I think the 
  nhmmer prior is more appropriate because it was trained on all
  columns of RNA alignments, whereas the singlet CM prior was trained
  only on single-stranded columns. Since the zero-bp models are meant
  for unstructured RNAs or RNAs of unknown structure, I think using
  the singlet prior, which has a strong A bias for example, is a bad
  idea. Ideally we'd have a prior trained on alignments of
  unstructured RNAs, but there's not enough data for that. 

o Models with zero basepairs need not be calibrated to be run in
  cmsearch and cmpress (and subsequently cmscan), because the CM 
  E-values are not used in the HMM-only pipeline variant.

o cmbuild: renamed --ignorant to --noss. This strips bps information from 
  SS_cons in input alignment, and, now, allows building models with no 
  SS_cons, they'll have 0 bps.

These changes were made with Rfam curators in mind. Their current
practice when faced with a RNA 'family' (often a single sequence) of
unknown structure is to predict one as a first step before building
the model. These changes should make it significantly easier for them
to build zero-bp models and do HMM-only searches first (since that
model need not be calibrated, and searches will be fast). Then they
can add a structure, but probably only if multiple homologs are found,
and an alignment of those homologs suggests a conserved structure.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3987 db3e70e4-12e1-0310-b23f-81a6e3d09ba7