More changes due to the introduction of the HMM-only pipeline in r3984.
Updated cmemit and cmstat so they use the filter HMM instead of the CM
when the CM has 0 basepairs, just like cmsearch and cmscan do. In
cmemit, sequences are emitted from the filter HMM instead of the
CM. In cmstat, E-value statistics from the filter HMM are output
instead of E-value statistics from the CM. --nohmmonly forces use of
the CM for all models in both programs.

Updated cmcalibrate so it doesn't calibrate CMs with 0 basepairs by
default, since cmsearch and cmscan won't use the CM. Such models are
skipped and an informative message is output explaining why they were
skipped. The CCOM (cmcalibrate command) annotation in the CM file is
updated, but CM E-value stats are not added to the CM
file. --nohmmonly forces calibration of all models.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3992 db3e70e4-12e1-0310-b23f-81a6e3d09ba7