Rethought and reverted changes in previous revisions to cmcalibrate (r3992) and cmpress (r3987). Briefly, cmcalibrate no longer skips calibration of zero basepair models by default, and cmpress requires all models in the CM file being pressed be calibrated even ones with zero basepairs. The rationale for these changes is discussed below.
As of r3992, models with zero basepairs would not be calibrated unless
--nohmmonly was used. The rationale was that an HMM would be used by
default in cmsearch/cmscan so CM E-value stats were
unnecessary. Similarly, as of r3987, cmpress would not require models
with zero basepairs be calibrated (all other models must be calibrated
though). This allowed cmscan to be run with a CM database in which
zero basepair CMs were not calibrated. This worked fine because HMM
E-value stats are calibrated at the cmbuild step, a la hmmbuild.

However, I realized that using -g (glocal mode) or --nohmmonly in
cmsearch or cmscan would require CM E-value stats even for models with
zero basepairs (since the HMM can't be run in glocal mode, so the CM
would need to be used). This meant that a user who wanted to use
glocal mode would have use --nohmmonly with cmcalibrate to force
calibration of zero bp models.

I don't want to require users know these caveats regarding 0 basepair
models, so I've reverted cmcalibrate to how it was prior to the r3992
commit (last prior change to cmcalibrate.c was r3968). That is,
cmcalibrate will always calibrate all models, even ones with 0
basepairs.  I've also reverted cmpress to require E-value stats for
all models.

It is still possible to use a non-calibrated zero basepair model in
cmsearch, but of course this will only happen for users that know they
can skip the calibration for these models (i.e. Rfam curators or those
rare, lovable users who actually read the documentation). I imagine
this ability to skip calibration could be useful when trying to build
a new family starting with one or a few sequences of unknown structure
where an initial fast HMM-only search can help you gain additional
homologs from which a structure might be inferred, as the Rfam
curators do (this situation is what motivated the development of the
HMM-only pipeline in the first place). However, before a release Rfam
would still want to calibrate the final model, even if it still has no
basepairs, because users who download the CM file should not be
limited. That is, they should be able to run glocal searches if they
want to.

For cmscan, all models need to be calibrated prior to running, since
cmpress requires calibrated models. I think this is reasonable
because cmscan should usually be run with a stable library of CMs,
e.g. a set of Rfam models from a specific release, which are probably
worth taking the time to calibrate, even if only to allow users to do
glocal cmscans.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3994 db3e70e4-12e1-0310-b23f-81a6e3d09ba7