Modified how RF annotation is transferred from the training alignment to a CM and from a CM to a cmalign output alignment, based on a recent conversation with Sean.
Now, if --hand is used in cmbuild msa->rf annotation is transferred to
the CM, else it isn't. Previously, training alignment RF annotation
was always transferred to the model if it existed.

cmalign now outputs #=GC RF annotation from the CM (cm->rf) if it
exists, else it uses the consensus sequence. Previously the consensus
sequence was always used. 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4023 db3e70e4-12e1-0310-b23f-81a6e3d09ba7