cm_mx.c, hmmband.c, infernal.h: Modified how the mode (JLRT) of truncated alignments to allow at each state is determined in rare cases where no CP9 HMM node has occupancy probability above 0.98 (the thresh2 paramter of a CP9Bands_t object).
Previously, in this case the sp2 and ep2 values of cm->cp9b
(CP9Bands_t) would be set to special values that had the effect that
all types of marginal alignment were possible at all states. This
results in a prohibitively large matrix for large models (up to 250Gb
for eukarya SSU).

As of this revision, if no HMM node has occ prob > 0.98 then we look
for the first and final HMM node that has occ prob > thresh2_failover,
(a new parameter in the CP9Bands_t object) which is 0.90. If no node
has at least this occ probability then we go back to old behavior of
allowing all marginal modes in all states.

There are 3 sequences out of the ~3 million in all Rfam 10.1 fulls for
which the old behavior resulted in DP matrix that exceeds the maximum
allowed default size (1028 Mb) in local alignment mode. All three are
in SSU_rRNA_eukarya and in all three cases the required matrix was way
above the limit (>100 Gb). The 'fix' implemented here with
thresh2_failover fixes all three cases such that an alignment can be
computed with a matrix < 1028 Mb.

Other changes:
cm_dpalign.c, cm_dpalign_trunc.c: changed an error message, only. 

cm_tophits.c: changed format of cyk score from %6.2f to %6.1f, output
only if --cyk.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4037 db3e70e4-12e1-0310-b23f-81a6e3d09ba7