Revisited and modified truncated DP alignment and search algorithms to allow the local end 'EL' state to be used in both L (Left) and R (Right) marginal modes. Previously, it was only possible to use EL in J (Joint) mode.
This fixes a design flaw in my truncated DP implementations. Diana
and Sean's algorithm (Bioinformatics, 2008) does not include ELs in L
or R mode but it doesn't have to since it allows a truncated alignment
to begin and end at any position in a target sequence. My
implementation requires that the first and/or final residue must be
included in a truncated parsetree. This is not only true in the
search/scan pipeline related to target sequence ends, but also in
cmalign where the full input sequence must be aligned, thus the first
and final residue must be included. So it is important that we can use
EL in L or R mode in the case that the first or final residue is best
aligned to the EL state.

Put another way, imagine sampling a parsetree and randomly truncating
it, if the chosen cut points occur within a stretch of EL emitted
residues our DP algorithms for inferring the parsetree given the
sequence ought to be able to score it that way - previous to this
revision they would not, but now they do.

The cmalign tests in ~/notebook/12_0423_inf_final_stress_tests/
brought my attention to this issue.

Also updated itest_brute.c to include paths that use EL in L or R
mode.

xrefs: 
/groups/eddy/home/nawrockie/notebook/12_0511_inf_trunc_LR_ELs/00LOG
/groups/eddy/home/nawrockie/notebook/12_0423_inf_final_stress_tests/00LOG


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4041 db3e70e4-12e1-0310-b23f-81a6e3d09ba7