Modified how HMM band tightening works in the rare case that the required HMM banded DP matrices exceed the maximum allowable size (1 Gb in cmalign).
Previously, the tau parameter which controls the amount of probability
mass outside the bands was doubled until it reached a maximum
value. Now, two additional parameters the CP9Bands_t object's thresh1
and thresh2 are modified as well. These parameters control the
occupancy probability required for a HMM position to be considered
'possibly' and 'probably' used by a parse and are used to define which
marginal alignment modes to use for each CM state in the downstream
truncated HMM banded alignment. This has the effect of further
reducing the required size of the DP matrices. 

Also, the maximum tau value has been raised to 0.05 from 0.01.

xref: ~/notebook/12_0423_inf_final_stress_tests/


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4042 db3e70e4-12e1-0310-b23f-81a6e3d09ba7