cm_alidisplay.c:
Modified PP annotation of EL emissions. Previously, these were left
blank (as ' '). Now, they are annotated with a single PP code
reporting the average PP of all emitted residues in the annotated
stretch. An example is shown below with an EL of 5 emissions
corresponding to a delete/skip of 7 model positions. The average
PP of the 5 EL emitted residues is between 0.75 and 0.85. 

          (((((((,,<<<<___.___.._>>>>,<<<<<~~~~~>>>>>,,,.,<<<<<_
  tRNA  1 GgggaugUAGCucAgU.GGU..AgaGCaucgGa*[7]*uCcgaagg.cgcgGGU
          G:GG:::UAG:::A+U G U  A:::C+:C:G:     :C:G:A + CG:GGGU
  test  1 GAGGUCUUAGUCUAAUuGCUcaAAGACGGCUGC*[5]*GCAGUAUAgCGUGGGU
          ****************9999************9..8..9***************

The dots in the PP line adjacent to the '8' don't actually correspond
to gaps in the alignment, but I couldn't come up with a clearly better
way of dealing with those columns. 

cm_dpsearch_trunc.c: 
Fixed an initialization bug recently introduced with L/R mode EL
states, caught by valgrind.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4043 db3e70e4-12e1-0310-b23f-81a6e3d09ba7