In the middle of fixing a problem with cmsearch -A <f> which outputs an alignment of all hits to file <f>. If a hit emits residues from the EL state these are not being annotated as '~' in the RF annotation of the output alignment. This is because cm_modelmaker.c:Transmogrify() does not handle EL emissions in a way that is consistent with how they are marked up in an output alignment created by cm_parsetree.c::Parsetrees2Alignment().
Transmogrify() has comments suggesting it can be used for local
alignments, but the local alignment conventions it expects are
different from what Parsetrees2Alignment() generates.  Transmogrify()
is (effectively) untouched since Infernal version 0.3, and I'm not
sure of the details of the local alignment conventions it expects
other than they are different from what Parsetree2Alignments()
generates (there's no further documentation I can find). It seems the
method of using these local alignments for building models was only
partially implemented. It doesn't much matter though because I've
verified that Transmogrify() is never currently being used with these
types of local alignments, at least in building all Rfam 10.1 models,
and they also are not used if the cmbuild input alignment was
generated by cmalign.

To avoid confusion with local alignments that use EL emissions, I've
reverted it to the Infernal v0.1 version of the function which did not
handle these other types of local alignments, and I plan to next add
support for the relevant types of local alignments I want the function
to be able to handle. The goal is to allow cmsearch -A to annotate EL
emissions in the RF annotation with '~' and also to allow cmbuild to
take an input alignment with '~' RF annotation and completely ignore
those columns (EL emissions are defined as 'nonhomologous'
residues). This way, cmbuild will work on any alignment generated by
cmalign and ignore ELs instead of treating them as insertions (which
it currently does). However, that change doesn't come in with this
revision; I'll work on that next.



Also, fixed a problem with annotation of PP for EL emissions in
cmsearch/cmscan hits introduced in revision 4043. For alignments that
use EL but without an EL emissions (silent local ends) the PP
annotation was being set as '0', now it's '.' for a gap, which makes
since there's no emitted residues.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4045 db3e70e4-12e1-0310-b23f-81a6e3d09ba7