Fixed problem with cmsearch -A output alignments that use EL emissions alluded to in log for r4045. Previously, EL emissions from hits were annotated as inserts. Now they are properly annotated as EL emissions with a '~' in the corresponding column of the RF annotation of the output alignment.
This fix required changing the HandModelmaker() and Transmogrify()
functions to optionally handle EL emissions for this case. The added
code is responsible for detecting the case where an EL has been used
for emissions from the aligned sequence. This is somewhat non-trivial
because EL emits are placed in the alignment in positions that can be
far away from the position modeled by the state that's transiting to
the EL. 

rnamat.c: Fixed misspelling of HandModelmaker() within a comment.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4049 db3e70e4-12e1-0310-b23f-81a6e3d09ba7